LPS-63567 don't abstract until we need to. Modules can still extend UpgradePortletRepository.java, and we don't know if future versions will need this class
